+++ tilellm/store/vector_store_repository.py
@@ -1,10 +1,14 @@
+import logging
 from abc import ABC, abstractmethod
 
 from langchain_core.vectorstores import VectorStore
 
 from tilellm.models.item_model import RepositoryNamespace, Engine, RepositoryItems, RepositoryNamespaceResult, \
     RepositoryDescNamespaceResult, QuestionAnswer
+from tilellm.tools.document_tools import get_content_by_url, load_document
 
+
+logger = logging.getLogger(__name__)
 
 class VectorStoreRepository(ABC):
 
@@ -120,3 +124,48 @@ class VectorStoreRepository(ABC):
         pass
 
 
+    @abstractmethod
+    async def initialize_embeddings_and_index(self, question_answer, llm_embeddings):
+        pass
+
+    @abstractmethod
+    async def perform_hybrid_search(self, question_answer, index, dense_vector, sparse_vector):
+        pass
+
+    @staticmethod
+    async def fetch_documents(type_source, source, scrape_type, parameters_scrape_type_4):
+        if type_source in ['url', 'txt']:
+            return await get_content_by_url(source,
+                                            scrape_type,
+                                            parameters_scrape_type_4=parameters_scrape_type_4)
+        return load_document(source, type_source)
+
+
+    @staticmethod
+    def calc_embedding_cost(texts, embedding):
+        """
+        Calculate the embedding cost with OpenAI embedding
+        :param texts:
+        :param embedding:
+        :return:
+        """
+        import tiktoken
+        enc = tiktoken.encoding_for_model('text-embedding-ada-002')
+        total_tokens = sum([len(enc.encode(page.page_content)) for page in texts])
+        logger.info(f'Total numer of Token: {total_tokens}')
+        cost = 0
+        try:
+            if embedding == "text-embedding-3-large":
+                cost = total_tokens / 1e6 * 0.13
+            elif embedding == "text-embedding-3-small":
+                cost = total_tokens / 1e6 * 0.02
+            else:
+                embedding = "text-embedding-ada-002"
+                cost = total_tokens / 1e6 * 0.10
+
+        except IndexError:
+            embedding = "text-embedding-ada-002"
+            cost = total_tokens / 1e6 * 0.10
+
+        logger.info(f'Embedding cost $: {cost:.6f}')
+        return total_tokens, cost